requirements.txt
fastapi
uvicorn
paddleocr
pillow
google-generativeai
